Improved logging to chef.log.

+++ recipes/ripgrep.rb
@@ -1,17 +1,23 @@
 # Install ripgrep.
+#
+# https://github.com/BurntSushi/ripgrep#installation
+# https://github.com/BurntSushi/ripgrep/releases
+#
+# See the Debian / Ubuntu section to install from the .deb file.
+
+return if skip_recipe
 
 bash "ripgrep" do
   code <<-EOT
-    exec >>~/chef.log 2>&1
-    chmod a+w ~/chef.log
-    echo -e "===\nLog ripgrep began `date`\n"
+    #{bash_began}
 
     rg --version >/dev/null 2>&1 || {
       curl -sL https://github.com/BurntSushi/ripgrep/releases/download/11.0.2/ripgrep_11.0.2_amd64.deb -o /tmp/ripgrep_11.0.2_amd64.deb
       dpkg -i /tmp/ripgrep_11.0.2_amd64.deb
       rm /tmp/ripgrep_11.0.2_amd64.deb
     }
+    echo "Say something else."
 
-    echo -e "\nLog ripgrep ended `date`"
+    #{bash_ended}
   EOT
 end
